requirements.txt
ultralytics
Flask
pillow
PyYAML